Added the ability to install TB4D, and get information from a web site
