+++ src/main/kotlin/com/balsdon/androidallyplugin/controller/AndroidStudioPluginController.kt
@@ -3,22 +3,28 @@ package com.balsdon.androidallyplugin.controller
 import com.android.ddmlib.AndroidDebugBridge
 import com.android.ddmlib.AndroidDebugBridge.IDeviceChangeListener
 import com.android.ddmlib.IDevice
-import com.android.ddmlib.IDevice.PROP_BUILD_API_LEVEL
-import com.android.ddmlib.IDevice.PROP_BUILD_VERSION
-import com.balsdon.androidallyplugin.extension.friendlyName
+import com.balsdon.androidallyplugin.TB4DInstallHelpWebPage
+import com.balsdon.androidallyplugin.TB4DWebPage
+import com.balsdon.androidallyplugin.localize
 import com.balsdon.androidallyplugin.model.AndroidDevice
 import com.balsdon.androidallyplugin.model.AndroidStudioPluginNotificationPayload
 import com.balsdon.androidallyplugin.model.NotificationPayload
 import com.balsdon.androidallyplugin.utils.log
+import com.intellij.ide.BrowserUtil
+import com.intellij.notification.Notification
+import com.intellij.notification.NotificationAction
 import com.intellij.notification.NotificationGroupManager
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.Project
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.subjects.BehaviorSubject
+import javax.swing.event.HyperlinkEvent
 
 
 class AndroidStudioPluginController(private val project: Project) : Controller {
     private val groupId = "AndroidAlly"
-    private val connectedDeviceList: MutableList<AndroidDevice.Device> = mutableListOf()
+    private val connectedDeviceList: MutableList<AndroidDevice> = mutableListOf()
     private val androidDeviceSource = BehaviorSubject.create<Set<AndroidDevice>>()
     override var selectedDeviceSerialList: MutableSet<String> = mutableSetOf()
 
@@ -37,26 +43,45 @@ class AndroidStudioPluginController(private val project: Project) : Controller {
             .notify(project)
     }
 
+    override fun showInstallTB4DSuccessNotification() {
+        showNotification(AndroidStudioPluginNotificationPayload(
+            localize("tb4d.install.success.title"),
+            localize("tb4d.install.success.message"),
+            actions = listOf(
+                object:NotificationAction(localize("tb4d.install.success.action")){
+                    override fun actionPerformed(event: AnActionEvent, notification: Notification) {
+                        BrowserUtil.browse(TB4DWebPage)
+                    }
+                }
+            )
+        ))
+    }
+
+    override fun showInstallTB4DErrorNotification() {
+        showNotification(AndroidStudioPluginNotificationPayload(
+            localize("tb4d.install.error.title"),
+            localize("tb4d.install.error.message"),
+            actions = listOf(
+                object:NotificationAction(localize("tb4d.install.error.action")){
+                    override fun actionPerformed(event: AnActionEvent, notification: Notification) {
+                        BrowserUtil.browse(TB4DInstallHelpWebPage)
+                    }
+                }
+            )
+        ))
+    }
+
     /* The guidance is to avoid initialisations in constructors
-     * @see [https://plugins.jetbrains.com/docs/intellij/plugin-extensions.html#implementing-extension]
-     *
-     * However, this is core functionality
-    */
+         * @see [https://plugins.jetbrains.com/docs/intellij/plugin-extensions.html#implementing-extension]
+         *
+         * However, this is core functionality
+        */
     init {
         AndroidDebugBridge.addDeviceChangeListener(object : IDeviceChangeListener {
             private fun updateDeviceList(device: IDevice, updateDevice: Boolean = true) {
                 connectedDeviceList.removeIf { it.serial == device.serialNumber }
                 if (updateDevice) {
-                    connectedDeviceList.add(
-                        AndroidDevice.Device(
-                            rawDevice = device,
-                            isEmulator = device.isEmulator,
-                            serial = device.serialNumber,
-                            apiLevel = device.getProperty(PROP_BUILD_API_LEVEL) ?: "",
-                            sdkLevel = device.getProperty(PROP_BUILD_VERSION) ?: "",
-                            friendlyName = device.friendlyName
-                        )
-                    )
+                    connectedDeviceList.add(AndroidDevice(device))
                 }
                 androidDeviceSource.onNext(connectedDeviceList.toSet())
             }
