+++ src/main/resources/Strings.properties
@@ -15,6 +15,7 @@ panel.device.no_devices=No devices
 panel.device.has_devices=Select a device
 
 panel.device.label.device_unknown=Unknown device
+panel.device.label.install.tb4d=Install TB4D
 
 android.accessibility.service.name.accessibility-menu=Accessibility menu
 android.accessibility.service.name.voice-access=Voice access
@@ -22,3 +23,11 @@ android.accessibility.service.name.talkback-for-developers=Talkback for develope
 android.accessibility.service.name.accessibility-scanner=Accessibility scanner
 android.accessibility.service.name.switch-access=Switch access
 android.accessibility.service.name.display-service=Display service
+
+tb4d.install.success.title=Success
+tb4d.install.success.message=TalkBack For Developers Installed
+tb4d.install.success.action=Learn more
+
+tb4d.install.error.title=Error
+tb4d.install.error.message=A problem occurred installing TalkBack For Developers
+tb4d.install.error.action=Open help
